Refactor trainer module
